Java source code
I have been running this in eclipse so you can just load it into eclipse
if that's what you want to use. I set the number of tabs to 6 but I
think 8 will be better so I'll fix that for my next commit. It will only
show your current access point unless you run it using sudo. To do so I
exported it as a runnable jar and ran 'sudo java -jar wifiscanner.jar'